refactor(entities/appointment): changing appointment-status enum folder
