+++ src/domain/entities/appointment/enum/appointment_status.enum.ts
