+++ test/factories/appointment_factory.ts
@@ -2,7 +2,7 @@ import { AppointmentEntity, AppointmentProps } from "@/domain/entities/appointme
 import { AppointmentComplaint } from "@/domain/entities/appointment/value-objects/complaint/appointment_complaint";
 import { AppointmentDateRange } from "@/domain/entities/appointment/value-objects/date-range/appointment_date_range";
 import { AppointmentSymptoms } from "@/domain/entities/appointment/value-objects/symptoms/appointment_symptoms";
-import { AppointmentStatusEnum } from "@/domain/models/enums/appointment_status.enum";
+import { AppointmentStatusEnum } from "@/domain/entities/appointment/enum/appointment_status.enum";
 import { EntityProps } from "@/domain/shared/entity";
 import { ONE_MILLISECOND } from "@/helpers/date_constants";
